Expose bustype parameter for Windows users
Prior to this commit using this project on a Windows machine
required editing the source code. After this commit the required
bustype configuration is exposed on the command line. This commit
also exposes the bitrate parameter and adds documentation to specify
the required usagages for both a Linux and Windows system.