bitrate parameter must be explicit int
Prior to this commit, using python3 on Windows, the CAN bus failed
to iniitalize because the CAN bitrate was not explicitly an int. This
resulted in an inability to use oscc-check. After this commit the
bitrate is explicitly set to an int type and the CAN bus is successfully
initialized on both Windows and Linux machines.