Add vehicle specification to Usage docs
Prior to this commit the usage sections of the README did not
address the vehicle specification option for the tool. This
could lead to users of the Kia Niro to see erronious reports
because the tool defaults to Kia Soul EV. After this
commit the example usages always specify the vehicle.