Incorporate Traivs CI into project
Prior to this commit there was no visibe indication on the project that it successfully built or that the tests were passing. This lead to an inablity to know if changes affected build or test status automatically. After this commit Travis CI has been connected to this project and there is a .travis.yml file that specifies how to set up and run the tests.