Typo correction in Readme and Docstring
Prior to this commit when you entered -h the docstring it output included a typo on line 4 in the help message for the -e option where checks was repeated twice.
I edited out the additional word in the dosctring as well as in the readme file that shows the docstring so that the additional word is not there.
This results in a message that is in proper and correct English.